% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_table_query.r
\docType{class}
\name{BlobStorage.JunctionTable.VirtualSQLProcedureCall-class}
\alias{BlobStorage.JunctionTable.VirtualSQLProcedureCall-class}
\title{VirtualBlobStorageSQLProcedureCall class}
\description{
Implements handling querries for BLOB storage
procedure Calls
}
\details{
Inherits from "VirtualSQLProcedureCall"
}

