% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_table_query.r
\docType{class}
\name{BlobStorage.SQLProcedureCall.ReferencedFileTable_SelectByParentTableName-class}
\alias{BlobStorage.SQLProcedureCall.ReferencedFileTable_SelectByParentTableName-class}
\title{BlobStorage.SQLProcedureCall.ReferencedFileTable_SelectByParentTableName class}
\description{
Implements handling querries for filetable name that is referenced by
associated junction table via path_locator foreign key
}
\details{
Inherits from "BlobStorage.VirtualSQLProcedureCall"
}

