% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetable_query.r
\docType{class}
\name{BlobStorage.SQLQuery.FileStoredInFileTable-class}
\alias{BlobStorage.SQLQuery.FileStoredInFileTable-class}
\title{Implements handling querries for files stored in filetables}
\description{
Inherits from "BlobStorage.VirtualSQLQuery"
}

