% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetable_query.r
\docType{class}
\name{BlobStorage.SQLQuery.FileTableRootPath-class}
\alias{BlobStorage.SQLQuery.FileTableRootPath-class}
\title{Implements handling querries for root path of file tables}
\description{
Inherits from "VirtualFileTableSQLQuerry"
}

