% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetable_storage_functions.r
\name{check_file_stored}
\alias{check_file_stored}
\title{check if file is stored in database}
\usage{
check_file_stored(filename, tb_name, db = .__DEFAULT_ODBC_DB_NAME__.,
  schema = .__DEFAULT_FILE_DB_SCHEMA__.)
}
\arguments{
\item{filename}{"character" filename of file to be stored in filetable}

\item{tb_name}{"character" name of the filetable}

\item{db}{"character" ODBC alias name of database}

\item{schema}{"character" name of DB schema}
}
\value{
\code{is_stored} "logical" TRUE if file is stored FALSE otherwise
}
\description{
check if file is stored in database
}

