% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetable_query.r
\name{generate_filetable_is_stored_query}
\alias{generate_filetable_is_stored_query}
\title{function to generate filetable query for stored file}
\usage{
generate_filetable_is_stored_query(keys)
}
\arguments{
\item{keys}{"data.frame" with keys}
}
\description{
function to generate filetable query for stored file
}

