% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetable_query.r
\name{generate_filetable_rootpath_query}
\alias{generate_filetable_rootpath_query}
\title{function to generate filetable rootpath query}
\usage{
generate_filetable_rootpath_query(keys)
}
\arguments{
\item{keys}{"data.frame" with keys}
}
\description{
function to generate filetable rootpath query
}

