% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blob_storage_functions.r
\name{get_filetable_path}
\alias{get_filetable_path}
\title{Get unc share path of file table}
\usage{
get_filetable_path(tb_name, db = .__DEFAULT_ODBC_DB_NAME__.,
  schema = .__DEFAULT_FILE_DB_SCHEMA__.)
}
\arguments{
\item{tb_name}{"character" name of the filetable}

\item{db}{"character" ODBC alias name of database}

\item{schema}{"character" name of DB schema}
}
\value{
\code{path} "character" unc path of the file table storage
}
\description{
Get unc share path of file table
}

