% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetable_storage_functions.r
\name{get_referenced_filetable_name}
\alias{get_referenced_filetable_name}
\title{Get name of file table associated with Junction Table}
\usage{
get_referenced_filetable_name(tb_name, db = .__DEFAULT_ODBC_DB_NAME__.,
  schema = .__DEFAULT_FILE_DB_SCHEMA__.)
}
\arguments{
\item{tb_name}{"character" name of the junction table}

\item{db}{"character" ODBC alias name of database}

\item{schema}{"character" name of DB schema}
}
\value{
\code{filetable} "character" name of the filetable
}
\description{
Get name of file table associated with Junction Table
}

