% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashing_functions.r
\name{hash_data_frame}
\alias{hash_data_frame}
\title{hash data frame}
\usage{
hash_data_frame(df, algo = "sha512")
}
\arguments{
\item{df}{"data.frame" to be hashed}

\item{algo}{"character" name of algorithm to be used as per "digest"
default is "sha512}
}
\value{
\code{ret} "character" hash of the data.frame
}
\description{
Hashes
}

