% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetable_query.r
\docType{methods}
\name{initialize,BlobStorage.SQLQuery.FileStoredInFileTable-method}
\alias{initialize,BlobStorage.SQLQuery.FileStoredInFileTable-method}
\title{Initialize method for "BlobStorage.SQLQuery.FileStoredInFileTable"}
\usage{
\S4method{initialize}{BlobStorage.SQLQuery.FileStoredInFileTable}(.Object,
  db_name, db_schema, tb_name, filename)
}
\arguments{
\item{.Object}{object of class derived from BlobStorage.SQLQuery.FileStoredInFileTable}

\item{db_name}{"character" database name}

\item{db_schema}{"character" database schema}

\item{tb_name}{"character" table name to be querried}

\item{filename}{"character" vector of filenames to be querried}
}
\description{
Initialize method for "BlobStorage.SQLQuery.FileStoredInFileTable"
}

