% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetable_query.r
\docType{methods}
\name{initialize,BlobStorage.SQLQuery.FileTableRootPath-method}
\alias{initialize,BlobStorage.SQLQuery.FileTableRootPath-method}
\title{Initialize method for "BlobStorage.SQLQuery.FileTableRootPath"}
\usage{
\S4method{initialize}{BlobStorage.SQLQuery.FileTableRootPath}(.Object, db_name,
  db_schema, tb_name)
}
\arguments{
\item{.Object}{object of class BlobStorage.SQLQuery.FileTableRootPath}

\item{db_name}{"character" database name}

\item{db_schema}{"character" database schema}

\item{tb_name}{"character" table name to be querried}
}
\description{
Initialize method for "BlobStorage.SQLQuery.FileTableRootPath"
}

