% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_table_query.r
\docType{methods}
\name{initialize,BlobStorage.VirtualSQLProcedureCall-method}
\alias{initialize,BlobStorage.VirtualSQLProcedureCall-method}
\title{Initialize method for "BlobStorage.VirtualSQLProcedureCall"}
\usage{
\S4method{initialize}{BlobStorage.VirtualSQLProcedureCall}(.Object, db_name,
  db_schema)
}
\arguments{
\item{.Object}{object of class derived from BlobStorage.VirtualSQLProcedureCall}

\item{db_name}{"character" database name}

\item{db_schema}{"character" database schema}
}
\description{
Initialize method for "BlobStorage.VirtualSQLProcedureCall"
}

