% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetable_query.r
\docType{methods}
\name{initialize,BlobStorage.VirtualSQLQuery-method}
\alias{initialize,BlobStorage.VirtualSQLQuery-method}
\title{Initialize method for "BlobStorage.VirtualSQLQuery"}
\usage{
\S4method{initialize}{BlobStorage.VirtualSQLQuery}(.Object, db_name, db_schema)
}
\arguments{
\item{.Object}{object of class derived from BlobStorage.VirtualSQLQuery}

\item{db_name}{"character" database name}

\item{db_schema}{"character" database schema}
}
\description{
Initialize method for "BlobStorage.VirtualSQLQuery"
}

