% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_table_query.r
\docType{methods}
\name{initialize,BlobStorage.SQLProcedureCall.JointFileTable_QueryByHashID-method}
\alias{initialize,BlobStorage.SQLProcedureCall.JointFileTable_QueryByHashID-method}
\title{Initialize method for "BlobStorage.SQLProcedureCall.JointFileTable_QueryByHashID"}
\usage{

  \S4method{initialize}{BlobStorage.SQLProcedureCall.JointFileTable_QueryByHashID}(.Object,
  db_name, db_schema, tb_name, keys = NULL)
}
\arguments{
\item{.Object}{object of class derived from FileTableSQLQuerry}

\item{db_name}{"character" database name}

\item{db_schema}{"character" database schema}

\item{tb_name}{"character" table name to be querried}

\item{keys}{"data.frame" with columns 'JointTableName' 'HashID'}
}
\description{
Initialize method for "BlobStorage.SQLProcedureCall.JointFileTable_QueryByHashID"
}

