% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_table_query.r
\docType{methods}
\name{initialize,BlobStorage.SQLProcedureCall.JointFileTable_QueryByTbNameTraderIDStartDateEndDate-method}
\alias{initialize,BlobStorage.SQLProcedureCall.JointFileTable_QueryByTbNameTraderIDStartDateEndDate-method}
\title{Initialize method for "SQLQuery.FileTableRootPath"}
\usage{

  \S4method{initialize}{BlobStorage.SQLProcedureCall.JointFileTable_QueryByTbNameTraderIDStartDateEndDate}(.Object,
  db_name, db_schema, tb_name, keys)
}
\arguments{
\item{.Object}{object of class derived from FileTableSQLQuerry}

\item{db_name}{"character" database name}

\item{db_schema}{"character" database schema}

\item{tb_name}{"character" table name to be querried}

\item{keys}{"data.frame" with columns 'TraderID', 'StartDate', 'EndDate'}
}
\description{
Initialize method for "SQLQuery.FileTableRootPath"
}

