% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_table_query.r
\docType{methods}
\name{initialize,BlobStorage.SQLProcedureCall.ReferencedFileTable_SelectByParentTableName-method}
\alias{initialize,BlobStorage.SQLProcedureCall.ReferencedFileTable_SelectByParentTableName-method}
\title{Initialize method for "SQLQuery.FileTableRootPath"}
\usage{

  \S4method{initialize}{BlobStorage.SQLProcedureCall.ReferencedFileTable_SelectByParentTableName}(.Object,
  db_name, db_schema, tb_name)
}
\arguments{
\item{.Object}{object of class derived from FileTableSQLQuerry}

\item{db_name}{"character" database name}

\item{db_schema}{"character" database schema}

\item{tb_name}{"character" table name to be querried}
}
\description{
Initialize method for "SQLQuery.FileTableRootPath"
}

