% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetable_query.r
\docType{methods}
\name{initialize,VirtualFileTableSQLQuery-method}
\alias{initialize,VirtualFileTableSQLQuery-method}
\title{Initialize method for "VirtualSQLQuery"}
\usage{
\S4method{initialize}{VirtualFileTableSQLQuery}(.Object, db_name, db_schema)
}
\arguments{
\item{.Object}{object of class derived from FileTableSQLQuerry}

\item{db_name}{"character" database name}

\item{db_schema}{"character" database schema}
}
\description{
Initialize method for "VirtualSQLQuery"
}

