% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_table_query.r
\docType{class}
\name{BlobStorage.SQLProcedureCall.JointFileTable_QueryByTbNameTraderIDStartDateEndDate-class}
\alias{BlobStorage.SQLProcedureCall.JointFileTable_QueryByTbNameTraderIDStartDateEndDate-class}
\title{BlobStorage.SQLProcedureCall.JointFileTable_QueryByTbNameTraderIDStartDateEndDate class}
\description{
Implements handling querries for joint table that
stores keys of associated file table. Returns file name column that
is associated with given key
}
\details{
Inherits from "BlobStorage.VirtualSQLProcedureCall"
}

