% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_table_query.r
\docType{class}
\name{BlobStorage.SQLProcedureCall.JointFileTable_UpdateByTbNameTraderIDInstrumentIDLegStartDateLegEndDate-class}
\alias{BlobStorage.SQLProcedureCall.JointFileTable_UpdateByTbNameTraderIDInstrumentIDLegStartDateLegEndDate-class}
\title{BlobStorage.SQLProcedureCall.JointFileTable_UpdateByTbNameTraderIDInstrumentIDLegStartDateLegEndDate class}
\description{
Implements handling update for joint table that
stores keys of associated file table. Returns number of stored and
insertet keys
}
\details{
Inherits from "BlobStorage.VirtualSQLProcedureCall"
}

