% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_table_query.r
\docType{class}
\name{BlobStorage.SQLProcedureCall.JointFileTable_UpdateByHashID-class}
\alias{BlobStorage.SQLProcedureCall.JointFileTable_UpdateByHashID-class}
\alias{BlobStorage.SQLProcedureCall.JointFileTable_UpdateByTbNameTraderIDStartDateEndDate-class}
\title{BlobStorage.SQLProcedureCall.JointFileTable_UpdateByHashID class}
\description{
Implements handling update for joint table that
stores keys of associated file table. Returns number of stored and
insertet keys

Implements handling update for joint table that
stores keys of associated file table. Returns number of stored and
insertet keys
}
\details{
Inherits from "BlobStorage.VirtualSQLProcedureCall"

Inherits from "BlobStorage.VirtualSQLProcedureCall"
}

