% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetable_storage_functions.r
\name{remove_file_from_filetable}
\alias{remove_file_from_filetable}
\title{Remove file from file table}
\usage{
remove_file_from_filetable(filename, tb_name, db = .__DEFAULT_ODBC_DB_NAME__.,
  schema = .__DEFAULT_FILE_DB_SCHEMA__.)
}
\arguments{
\item{filename}{"character" filename of  file to be removed from filetable}

\item{tb_name}{"character" name of the filetable}

\item{db}{"character" ODBC alias name of database}

\item{schema}{"character" name of DB schema}
}
\value{
\code{status} "integer" status of saving:\cr
 \code{0} - file has been removed\cr
 \code{-1} - file has not been stored and couldn't be removed
}
\description{
Remove file from file table
}

