% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetable_query.r
\docType{class}
\name{SQLQuery.FileStoredInFileTable-class}
\alias{SQLQuery.FileStoredInFileTable-class}
\title{SQLQuery.FileTableRootPath class}
\description{
Implements handling querries for root path of file tables
}
\details{
Inherits from "VirtualFileTableSQLQuerry"
}

