% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetable_storage_functions.r
\name{store_file_in_filetable}
\alias{store_file_in_filetable}
\title{Store file in file table}
\usage{
store_file_in_filetable(file, tb_name, db = .__DEFAULT_ODBC_DB_NAME__.,
  schema = .__DEFAULT_FILE_DB_SCHEMA__., overwrite = FALSE)
}
\arguments{
\item{file}{"character" full file path of the file to be stored in filetable}

\item{tb_name}{"character" name of the filetable}

\item{db}{"character" ODBC alias name of database}

\item{schema}{"character" name of DB schema}

\item{overwrite}{"logical" should file be overwritten if already exists?}
}
\value{
\code{status} "integer" status of saving:\cr
 \code{0} - new file has been stored\cr
 \code{1} - file has been overwritten\cr
 \code{-1} - file has not been stored as it already exists and \code{overwrite} param was set to FALSE
}
\description{
Store file in file table
}

