% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetable_query.r
\docType{class}
\name{VirtualFileTableSQLQuery-class}
\alias{VirtualFileTableSQLQuery-class}
\title{VirtualFileTableSQLQuery class}
\description{
Implements handling querries for File Tables
}
\details{
Inherits from "VirtualSQLQuery"
}

