% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junction_table_query.r
\docType{class}
\name{BlobStorage.SQLProcedureCall.JointFileTable_UpdateByHashID-class}
\alias{BlobStorage.SQLProcedureCall.JointFileTable_UpdateByHashID-class}
\alias{BlobStorage.SQLProcedureCall.JointFileTable_UpdateByTbNameTraderIDStartDateEndDate-class}
\title{BlobStorage.SQLProcedureCall.JointFileTable_UpdateByHashID class}
\description{
Implements handling querries for joint table that
stores keys of associated file table. Returns file name column that
is associated with given key

Implements handling querries for joint table that
stores keys of associated file table. Returns file name column that
is associated with given key
}
\details{
Inherits from "BlobStorage.VirtualSQLProcedureCall"

Inherits from "BlobStorage.VirtualSQLProcedureCall"
}

